Update 5-relationships-and-hyperlinked-apis.md
Following  [@jpadilla](https://github.com/jpadilla)'s suggestion in issue [#2837](https://github.com/tomchristie/django-rest-framework/issues/2837), I am proposing a little addition in "Creating an endpoint for the root of our API" section in order to make it clear that the `user-list` and `snippet-list` names will be declared later in `snippets/urls.py`. Missing this clarification might lead a beginner (like me) to think that the names are somehow assigned by rest_framework under the hood and to try the code too early, getting stuck in the following errors.